IMU platform/requirements.txt
asgiref==3.4.1
cycler==0.10.0
Django==3.2.7
django-multiselectfield==0.1.12
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
